Renomme Checkbox en Checkboxes pour faciliter l'usage de Checkbox
