+++ src/components/form/dsfr/Checkbox.tsx
@@ -1,42 +1,11 @@
-import DsfrCheckbox, { type CheckboxProps as DsfrCheckboxProps } from '@codegouvfr/react-dsfr/Checkbox';
-import styled, { css } from 'styled-components';
+import Checkboxes, { type CheckboxesProps } from './Checkboxes';
 
-const StyledCheckbox = styled(DsfrCheckbox)<{
-  small?: boolean;
-  $fullWidth?: boolean;
-}>`
-  ${({ small, $fullWidth }) => css`
-    ${small &&
-    css`
-      .fr-fieldset__content .fr-radio-group label,
-      .fr-fieldset__content .fr-checkbox-group label {
-        padding: 0.25rem 0;
-      }
-      .fr-fieldset__content .fr-radio-group:not(.fr-radio-rich) input[type='radio'] + label::before,
-      .fr-fieldset__content .fr-checkbox-group input[type='checkbox'] + label::before {
-        top: 0.25rem;
-      }
-      .fr-label {
-        font-size: 0.875rem;
-      }
-    `}
-    ${!$fullWidth &&
-    css`
-      .fr-label {
-        display: inline-flex !important; /* Prevent label to be clickable on the whole line even when label is small */
-      }
-    `}
-  `}
-`;
+type CheckboxOption = CheckboxesProps['options'][number];
+export type CheckboxProps = Omit<CheckboxesProps, 'options'> &
+  Omit<CheckboxOption, 'nativeInputProps'> & { nativeInputProps?: CheckboxOption['nativeInputProps'] & { name: string } };
 
-export type CheckboxProps = DsfrCheckboxProps & {
-  fullWidth?: boolean;
-};
-
-const Checkbox: React.FC<CheckboxProps> = ({ fullWidth = true, ...props }) => {
-  // HACK to force the checkbox to be re-rendered when the options change from outside
-  const key = props.options.map(({ nativeInputProps: { value } }) => value).join('_');
-  return <StyledCheckbox key={key} $fullWidth={fullWidth} {...props} />;
+const Checkbox: React.FC<CheckboxProps> = ({ label, hintText, nativeInputProps, ...props }) => {
+  return <Checkboxes options={[{ label, hintText, nativeInputProps: nativeInputProps || {} }]} {...props} />;
 };
 
 export default Checkbox;
